weapon stuff changed to weapon component, control moved to weapon control, removed weapon system.
Former-commit-id: b3b2716804f3af2f22b9e388bf578779bac8a84b

+++ EngineComponentSystem/Entities/IEntity.cs
@@ -17,7 +17,7 @@ namespace Engine.ComponentSystem.Entities
         /// <summary>
         /// A globally unique id for this object.
         /// </summary>
-        long UID { get; set; }
+        int UID { get; set; }
 
         /// <summary>
         /// A list of all of this entities components.
